% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_measurement_schedule.R
\name{compute_measurement_schedule}
\alias{compute_measurement_schedule}
\title{Scales times values according to the designated time period (new comment).}
\usage{
compute_measurement_schedule(
  time_period,
  num_measurements,
  base_time_length,
  measurement_spacing
)
}
\arguments{
\item{time_period}{number of days over which change occurs}

\item{num_time_points}{number of time points}
}
\value{
Returns a data table.
}
\description{
Scales times values according to the designated time period (new comment).
}
