% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cell_simulation.R
\name{compute_mult_simulations}
\alias{compute_mult_simulations}
\title{Scales times values according to the designated time period (new comment).}
\usage{
compute_mult_simulations(
  num_sims,
  pop_params,
  cov_matrix,
  schedule,
  starting_values,
  response_group_size
)
}
\arguments{
\item{time_period}{number of days over which change occurs}

\item{num_time_points}{number of time points}
}
\value{
Returns a data table.
}
\description{
Scales times values according to the designated time period (new comment).
}
