% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_starting_values_4l.R
\name{compute_starting_values_4l}
\alias{compute_starting_values_4l}
\title{Computes starting values for four-parameter logistic curve model.}
\usage{
compute_starting_values_4l(data)
}
\arguments{
\item{data_wide}{wide version of data}

\item{model_name}{name of model}
}
\description{
Data are modelled with a structured latent growth curve model.
}
