% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\name{create_logistic_growth_model}
\alias{create_logistic_growth_model}
\title{Analyzes data with nonlinear latent growth curve model.}
\usage{
create_logistic_growth_model(
  data,
  model_name,
  diff_fixed_est,
  beta_fixed_est,
  gamma_fixed_est,
  var_diff_est,
  var_beta_est,
  var_gamma_est,
  error
)
}
\arguments{
\item{data}{number of measurements}

\item{param_table}{table containing parameter values for each person}

\item{measurement_days}{days of measurement}

\item{time_period}{length of time over which measurements are taken}
}
\value{
Returns a data table.
}
\description{
Description of function.
}
