% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_logistic_growth_model_4l.R
\name{create_logistic_growth_model_4l}
\alias{create_logistic_growth_model_4l}
\title{Analyzes data with nonlinear latent growth curve model.}
\usage{
create_logistic_growth_model_4l(data_wide, model_name, starting_values)
}
\arguments{
\item{data_wide}{wide version of data}

\item{model_name}{name of model}
}
\description{
Data are modelled with a structured latent growth curve model.
}
