% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_convergence.R
\name{evaluate_convergence}
\alias{evaluate_convergence}
\title{Tests effectiveness of starting value procedure.}
\usage{
evaluate_convergence(
  factor_list,
  num_iterations,
  pop_params,
  response_group_size,
  num_cores
)
}
\arguments{
\item{factor_list}{}

\item{num_iterations}{number of time points}

\item{pop_params}{}
}
\value{
Returns a data table
}
\description{
Tests effectiveness of starting value procedure.
}
