% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_models.R
\name{evaluate_models}
\alias{evaluate_models}
\title{Fits model with several sets of starting values.}
\usage{
evaluate_models(initial_model, starting_values)
}
\arguments{
\item{data}{number of measurements}

\item{param_table}{table containing parameter values for each person}

\item{measurement_days}{days of measurement}

\item{time_period}{length of time over which measurements are taken}
}
\value{
Returns a data table.
}
\description{
Description of function.
}
