% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_best_estimates.R
\name{extract_best_estimates}
\alias{extract_best_estimates}
\title{Extracts parameter estimates from best-fitting model}
\usage{
extract_best_estimates(fit_results, starting_values)
}
\arguments{
\item{data}{number of measurements}

\item{param_table}{table containing parameter values for each person}

\item{measurement_days}{days of measurement}

\item{time_period}{length of time over which measurements are taken}
}
\value{
Returns a data table.
}
\description{
Description of function.
}
