% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_person_scores_4l.R
\name{generate_group_scores_4l}
\alias{generate_group_scores_4l}
\title{Generates scores at each time point for each person.}
\usage{
generate_group_scores_4l(
  num_measurements,
  param_table,
  measurement_days,
  time_period
)
}
\arguments{
\item{num_measurements}{number of measurements}

\item{param_table}{table containing parameter values for each person}

\item{measurement_days}{days of measurement}

\item{time_period}{length of time over which measurements are taken}
}
\value{
Returns a data table.
}
\description{
Scores are generates according to three-parameter logistic function.
}
