% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_start_values.R
\name{generate_multiple_starting_value_sets}
\alias{generate_multiple_starting_value_sets}
\title{Generates starting values for estimating nonlinear parameters.}
\usage{
generate_multiple_starting_value_sets(num_sets, data, data_wide)
}
\arguments{
\item{num_sets}{number of starting value sets to generate}

\item{time_fixed_effects}{names of fixed-effects parameters relating to time (i.e., beta and gamma in case of three-parameter logistic curve)}

\item{time_rand_effects}{names of random-effects parameters relating to time (i.e., beta and gamma in case of three-parameter logistic curve)}

\item{diff_effects}{names of parameters that define the difference between starting and ending values of the logistic curve}

\item{error}{name of error parameter}
}
\value{
Returns a data table of starting values
}
\description{
Starting value procedure largely mirrors that of MPLUS.
}
