% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_exp_simulations_4l.R
\name{run_exp_simulation_4l}
\alias{run_exp_simulation_4l}
\title{Runs Monte Carlo simulations.}
\usage{
run_exp_simulation_4l(
  factor_list,
  num_iterations,
  pop_params,
  response_group_size,
  num_cores,
  seed
)
}
\arguments{
\item{num_iterations}{number of iterations}

\item{pop_params}{list of population parameters}

\item{response_group_size}{sample size}

\item{schedule}{measurement schedule}
}
\value{
Returns a data table.
}
\description{
Runs Monte Carlo simulations.
}
