% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulations.R
\name{run_simulations}
\alias{run_simulations}
\title{Runs Monte Carlo simulations.}
\usage{
run_simulations(
  num_iterations,
  pop_params,
  cov_matrix,
  schedule,
  response_group_size
)
}
\arguments{
\item{num_iterations}{number of iterations}

\item{pop_params}{list of population parameters}

\item{schedule}{measurement schedule}

\item{response_group_size}{sample size}
}
\value{
Returns a data table.
}
\description{
Runs Monte Carlo simulations.
}
